Added a few recipes to Nacrotic bone to the diagram
Putting a bone in the coke oven. (pretty early game, but takes a long time)
Putting bone and ash in the alloy smelter. (faster than the coke oven but also requires ash)
Putting bone, ash, fermented spider eye, and coal into the pressure chamber (possibly the fastest but requires a lot of setup and more resources)
Putting bone, ash, fermented spider eye, and coal dust into the portal of alfheim (as fast as the pressure chamber but using botania instead if that's your style)
close #111